chore: refactor naming convention
